Update README.md
Craig, I took the liberty of taking the email you sent out last week to add to your README file here. That way it'll be easier to point people here in the future w/o having to compose that documentation again. To make it easiest on you, I'm submitting my edits as a Pull Request.

In the future, we can think about creating a climatology methods repo on the NANOOS github organization, to be based mostly on your stuff; at least initially.